refactor(bank): simplify architecture — four cohesive changes
Four independent simplifications applied together because the file
renames (nodes/ promotion) mean intermediate states don't pass
mypy/ruff, so they can't be committed separately.

1. Extract parse_ofx_file() shared helper; collapse OFX handlers
   apple_card_ofx.py and apple_savings_ofx.py were ~75% identical.
   parse() and _parse_transactions() were 100% identical (67 lines).
   Extracted parse_ofx_file(file_path, include_available) to base.py.
   Each OFX handler collapses from ~125 lines to ~30 lines. The
   include_available flag captures credit-card vs. savings distinction
   without inheritance. Removed stale S314 ruff suppressions.

2. Extract parse_ynab_import_posted_date() to matching.py
   YNAB import_id date parsing was an inline regex loop in flow.py.
   Moved to matching.py alongside YNAB_PAYEE_EXPANSIONS and
   make_import_id() where all other YNAB-specific matching knowledge
   lives. Removed import re from flow.py.

3. Promote nodes/ subpackage to top-level modules
   nodes/ contained thin wrappers (apply.py, parse.py, reconcile.py,
   retrieve.py) that added an indirection layer without decoupling
   benefit. Promoted to bank_accounts/ top level via git mv. Deleted
   nodes/__init__.py. Updated all import paths in flow.py, test files,
   and pyproject.toml ruff config.

4. Merge double unmatched-YNAB-tx iteration into one pass
   reconcile_account_data() iterated unmatched_ynab_txs twice:
   once to generate DeleteOps, once to categorize with mismatch_reason.
   Merged into a single pass computing _classify_mismatch_reason once.
