+++ pyproject.toml
@@ -148,9 +148,7 @@ ignore = [
 "tests/fixtures/synthetic_data.py" = ["S311"]  # allow standard random for test data generation
 "tests/e2e/**/*" = ["S105", "S607"]  # allow test tokens and partial executable paths for CLI testing
 "src/finances/ynab/flow.py" = ["S607"]  # allow partial executable path for ynab CLI tool
-"src/finances/bank_accounts/nodes/apply.py" = ["S603", "S607"]  # allow subprocess calls to ynab CLI tool
-"src/finances/bank_accounts/format_handlers/apple_card_ofx.py" = ["S314"]  # OFX files are trusted bank exports, not untrusted user input
-"src/finances/bank_accounts/format_handlers/apple_savings_ofx.py" = ["S314"]  # OFX files are trusted bank exports, not untrusted user input
+"src/finances/bank_accounts/apply.py" = ["S603", "S607"]  # allow subprocess calls to ynab CLI tool
 
 # MyPy configuration
 [tool.mypy]
