+++ tests/integration/test_bank_accounts/test_reconcile_node.py
@@ -12,9 +12,8 @@ from finances.bank_accounts.models import (
     BankTransaction,
     ImportPattern,
 )
-from finances.bank_accounts.nodes import reconcile_account_data
-from finances.bank_accounts.nodes.reconcile import YNAB_STARTING_BALANCE_PAYEE
 from finances.bank_accounts.operations import CreateOp, DeleteOp, FlagOp
+from finances.bank_accounts.reconcile import YNAB_STARTING_BALANCE_PAYEE, reconcile_account_data
 from finances.core import FinancialDate, Money
 from finances.core.json_utils import write_json
