+++ tests/unit/test_bank_accounts/test_reconcile_node.py
@@ -10,8 +10,8 @@ from finances.bank_accounts.models import (
     BankAccountsConfig,
     ImportPattern,
 )
-from finances.bank_accounts.nodes.reconcile import _classify_mismatch_reason, reconcile_account_data
 from finances.bank_accounts.operations import CreateOp
+from finances.bank_accounts.reconcile import _classify_mismatch_reason, reconcile_account_data
 from finances.core import FinancialDate, Money
